+++ Sources/NSAttributedStringBuilder/SpellingState.swift
@@ -2,7 +2,7 @@
 public extension NSAttributedStringBuilder {
 
     /// Defines a spelling state for the attributed string. (macOS only)
-    enum SpellingState: Int {
+    enum SpellingState: Int, Hashable, Codable, Sendable {
         /// Indicates that the attributed string has no spelling errors.
         case noGrammarOrSpellingIssues
         /// Flag for spelling issues.