+++ Sources/NSAttributedStringBuilder/WritingDirection.swift
@@ -4,7 +4,7 @@ public extension NSAttributedStringBuilder {
 
     /// Combines `NSWritingDirection` and `NSWritingDirectionFormatType`
     /// into one type that can be used with `writingDirection(_:)`.
-    enum WritingDirection: NSNumber {
+    enum WritingDirection: NSNumber, Hashable, Codable, Sendable {
         case leftToRightDirectionEmbedding
         case rightToLeftDirectionEmbedding
         case leftToRightDirectionOverride